Translate {Field,Scalar}IfStmt to one MaskStmt
When lowering gtir to oir, translate `FieldIfStmt` and `SclarIfStmt`
to only one `MaskStmt`, avoiding the previous extra `AssignStmt`.

This is preporatory work for exposing control flow in the gt4py/dace
bridge.

We'll need to clean this up since this is - mostl likely - breaking some
assumptions later on. One way forward (keeping the `AssignStmt`) would
be to save it as part of the `MaskStmt` keeping the old structure but
making sure that `AssignStmt`  and the rest of the `MaskStmt`  "travel"
together.