This seems to generate something reasonable (prob wrong though)
Warning: the generated sdfg doesn't validate. However, validation
doesn't seem to be a requirement at the moment in orchestration. This is
weird and we should be careful.

That said, sdfg simplification runs through and the generated sdfg looks
okay although overcomplicated. The generated code shares the
overcomplications of the sdfg, looks valid at first glance, but has
nasty shadowing (dace:cpu) that (imo) will result in undefined behavior.