Avoid circular import by moving SDFGContext
Move the class to a separate file to avoid the circular import (which
emerged by adding type hints - see previous commit).

This whole commit might not be needed since SDFGContext is only used
once (in visit_IndexAccess and passed on to self.visit(idx, ...)). No
other visitor in Tasklet_Codegen is using it, so we might just be able
to get rid of it and move the class back to SDFGBuilder (where it was
extracted from).