WIP: ignore k offset write book keeping
not needed anymore since we now properly separate read and write memlets

not working either. needs fixes to how we setup the "conditional
evaluation tasklet", i.e. we don't go through the "normal" ComputeState
transformations and thus lack the updated node-ctx information. This is
why it tries to connect to outer maps.