WIP: this fails because of duplicate connectors
if input and output connectors are called the same, e.g. in code like

```python
level = 1
condition = True
while condition:
	condition = level < 10
	level = level + 1
```

which translates to

```none
Tasklet1:
	level = 1
	condtion = True

While condition:
	Tasklet2:
		condition = level < 10
		level = level + 1
```

then the sdfg complains because DaCe can't handle reading level and
writing to it in the same tasklet. This creates and input connector for
level and an output connector for level, which is ambiguous.